Add liveness.cpp/hpp logic and force offset resolution to be done early
We do this by forcibly including the il2cpp source directly. This required some hackiness to get around the bs-hook typedefs.h issues,
but ultimately this should be an improvement, since that source is relatively stable across Unity updates so far.

This should also help prevent many of spurious liveness issues that were plaguing people in the past.

Atomic logic and HasParentUnsafe logic taken from: https://github.com/QuestPackageManager/Il2CppQuestTypePatching/pull/24
