+++ qpm.shared.json
@@ -103,12 +103,12 @@
     {
       "dependency": {
         "id": "paper2_scotland2",
-        "versionRange": "=4.7.0",
+        "versionRange": "=4.8.0",
         "additionalData": {
-          "soLink": "https://github.com/Fernthedev/paperlog/releases/download/v4.7.0/libpaper2_scotland2.so",
+          "soLink": "https://github.com/Fernthedev/paperlog/releases/download/v4.8.0/libpaper2_scotland2.so",
           "overrideSoName": "libpaper2_scotland2.so",
-          "modLink": "https://github.com/Fernthedev/paperlog/releases/download/v4.7.0/paper2_scotland2.qmod",
-          "branchName": "version/v4_7_0",
+          "modLink": "https://github.com/Fernthedev/paperlog/releases/download/v4.8.0/paper2_scotland2.qmod",
+          "branchName": "version/v4_8_0",
           "compileOptions": {
             "systemIncludes": [
               "shared/utfcpp/source"
@@ -117,7 +117,7 @@
           "cmake": false
         }
       },
-      "version": "4.7.0"
+      "version": "4.8.0"
     },
     {
       "dependency": {
@@ -153,13 +153,13 @@
     {
       "dependency": {
         "id": "beatsaber-hook",
-        "versionRange": "=8.0.0",
+        "versionRange": "=8.1.0",
         "additionalData": {
-          "soLink": "https://github.com/QuestPackageManager/beatsaber-hook/releases/download/v8.0.0/libbeatsaber-hook.so",
-          "debugSoLink": "https://github.com/QuestPackageManager/beatsaber-hook/releases/download/v8.0.0/debug_libbeatsaber-hook.so",
+          "soLink": "https://github.com/QuestPackageManager/beatsaber-hook/releases/download/v8.1.0/libbeatsaber-hook.so",
+          "debugSoLink": "https://github.com/QuestPackageManager/beatsaber-hook/releases/download/v8.1.0/debug_libbeatsaber-hook.so",
           "overrideSoName": "libbeatsaber-hook.so",
-          "modLink": "https://github.com/QuestPackageManager/beatsaber-hook/releases/download/v8.0.0/beatsaber-hook.qmod",
-          "branchName": "version/v8_0_0",
+          "modLink": "https://github.com/QuestPackageManager/beatsaber-hook/releases/download/v8.1.0/beatsaber-hook.qmod",
+          "branchName": "version/v8_1_0",
           "compileOptions": {
             "cppFlags": [
               "-Wno-extra-qualification"
@@ -168,7 +168,7 @@
           "cmake": true
         }
       },
-      "version": "8.0.0"
+      "version": "8.1.0"
     },
     {
       "dependency": {
